fix : 메인 렌딩페이지 추가 및 리팩토링
