+++ package.json
@@ -18,6 +18,7 @@
     "@mui/icons-material": "^6.0.0",
     "@mui/material": "^6.0.0",
     "@tanstack/react-query": "^5.53.3",
+    "@tanstack/react-query-devtools": "^5.56.0",
     "@tiptap/extension-character-count": "^2.6.6",
     "@tiptap/extension-image": "^2.6.6",
     "@tiptap/extension-link": "^2.6.6",
